requirements.txt
asgiref==3.11.1
astroid==4.0.4
certifi==2026.2.25
charset-normalizer==3.4.4
datetime==6.0
dill==0.4.1
django==6.0.2
django-extensions==4.1
django-filter==25.2
django-redis==6.0.0
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
docker==7.1.0
gprof2dot==2025.4.14
hiredis==3.3.0
idna==3.11
isort==8.0.0
markdown==3.10.2
mccabe==0.7.0
pillow==12.1.1
pip==25.3
platformdirs==4.9.2
pyjwt==2.11.0
pylint==4.0.5
pytz==2026.1.post1
redis==7.2.1
requests==2.32.5
sqlparse==0.5.5
tomlkit==0.14.0
tzdata==2025.3
urllib3==2.6.3
uv==0.10.8
zope-interface==8.2
